Replace set with map (#3)
* feat: map

Now we have an id with each state, managing states with map can be more powerful
than set. Also hashing pointers as keys doesn't seem to be a good
choice.

* feat: support iteration on map

* test: map

* fix: to_next skipping the first slot

* refactor: delete NFA with map

* refactor: collect moves with map

* refactor: remove set from source

* feat: getter for size of map

* test: size of moves & closures should be correct

* fix: invalid read error

We were to use a post-order traversal on the NFA to delete all of the
states. But if there's a common state that two states transit to, the
second time becomes an invalid read when we look up the id.

* refactor: remove low-level detail & rename function